Abstraction of ViewHolder to prevent code repetition
