+++ android/src/main/java/nl/fhict/happynews/android/viewholder/PostImageHolder.java
@@ -18,9 +18,8 @@ import java.util.Date;
 /**
  * Created by tom on 27/03/2017.
  */
-public class PostImageHolder extends RecyclerView.ViewHolder implements View.OnClickListener {
+public class PostImageHolder extends ViewHolder{
 
-    private Post post;
     private TextView sourceTextView;
     private TextView timeTextView;
     private TextView headlineTextView;
@@ -34,51 +33,18 @@ public class PostImageHolder extends RecyclerView.ViewHolder implements View.OnC
         timeTextView = (TextView) view.findViewById(R.id.timeTextView);
         headlineTextView = (TextView) view.findViewById(R.id.headlineTextView);
         imageView = (ImageView) view.findViewById(R.id.imageView);
-
-        view.setOnClickListener(this);
     }
 
     public void bindType(Post post) {
-        this.post = post;
+        super.bindType(post);
 
-        sourceTextView.setText(this.post.getSource());
-        timeTextView.setText(relativeTimeSpan(this.post.getPublishedAt()));
-        headlineTextView.setText(this.post.getTitle());
+        sourceTextView.setText(post.getSource());
+        timeTextView.setText(relativeTimeSpan(post.getPublishedAt()));
+        headlineTextView.setText(post.getTitle());
 
         Ion.with(imageView)
-                .load(this.post.getImageUrls().get(0));
+                .load(post.getImageUrls().get(0));
     }
 
-    /**
-     * Creates a neatly formatted string displaying the Relative Time Span
-     *
-     * @param input the start time as a Date object
-     * @return relative Timestamp (eg. '4 hours ago')
-     */
-    private String relativeTimeSpan(Date input) {
-
-        long unixTime = input.getTime();
-
-        if (unixTime == 0) {
-            return DateUtils.getRelativeTimeSpanString(System.currentTimeMillis()).toString();
-        }
-
-        return DateUtils.getRelativeTimeSpanString(unixTime).toString();
-    }
-
-    @Override
-    public void onClick(View v) {
-
-        Context context = itemView.getContext();
-
-        if (!post.getUrl().isEmpty()) {
-            Uri uri = Uri.parse(post.getUrl());
-            CustomTabsIntent customTabsIntent = new CustomTabsIntent.Builder().build();
-            customTabsIntent.launchUrl(context, uri);
-        } else {
-            Toast.makeText(context, "Link not found",
-                    Toast.LENGTH_SHORT).show();
-        }
-    }
 }
